Szymon Kotłownia works on Linux (poorly tho)

+++ Xenon/src/devTools/logger.hpp
@@ -131,7 +131,7 @@ namespace Core {
 		{
 			const size_t tlc = findToken(text);
 			if (tlc == std::string::npos) { m_msg << text; return; }
-			m_msg.write(text, tlc);	
+			m_msg.write(text, static_cast<long>(tlc));	
 			proccesToken(text[tlc + 1], arg);
 			output(&text[tlc+3], args...);
 		}
